Move questPlay files
