+++ src/ui/questPlay/questPlay.colors.ts
@@ -1,4 +1,4 @@
-export const style = {
+export const colors = {
   textColor: "#9eacac",
   textColorInactive: "#747e80",
   textColorCrl: "#ffd863",
